# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "MiniBlog API",
    "version": "1.0.0",
    "description": "API REST para gestionar usuarios y posts"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "summary": "Obtener todos los posts",
        "tags": [
          "Posts"
        ],
        "responses": {
          "200": {
            "description": "Lista de posts"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo post",
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "author_id"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "author_id": {
                    "type": "integer"
                  },
                  "published": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Obtener un post por ID",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del post"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un post",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "published": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post actualizado"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un post",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post eliminado"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      }
    },
    "/posts/author/{authorId}": {
      "get": {
        "summary": "Obtener posts de un autor",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "authorId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts del autor"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Obtener todos los autores",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "Lista de autores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "bio": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo autor",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Autor creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Obtener un autor por ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del autor"
          },
          "404": {
            "description": "Autor no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un autor",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autor actualizado"
          },
          "404": {
            "description": "Autor no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un autor",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autor eliminado"
          },
          "404": {
            "description": "Autor no encontrado"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}